+++ src/Pim/Bundle/CatalogBundle/Builder/ProductBuilder.php
@@ -7,6 +7,7 @@ use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductPriceInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
+use Pim\Bundle\CatalogBundle\Repository\AssociationTypeRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\CurrencyRepositoryInterface;
@@ -33,6 +34,9 @@ class ProductBuilder implements ProductBuilderInterface
     /** @var string */
     protected $productPriceClass;
 
+    /** @var string */
+    protected $associationClass;
+
     /** @var AttributeRepositoryInterface */
     protected $attributeRepository;
 
@@ -48,16 +52,20 @@ class ProductBuilder implements ProductBuilderInterface
     /** @var CurrencyRepositoryInterface */
     protected $currencyRepository;
 
+    /** @var AssociationTypeRepositoryInterface */
+    protected $assocTypeRepository;
+
     /**
      * Constructor
      *
-     * @param AttributeRepositoryInterface $attributeRepository Attribute repository
-     * @param FamilyRepositoryInterface    $familyRepository    Family repository
-     * @param ChannelRepositoryInterface   $channelRepository   Channel repository
-     * @param LocaleRepositoryInterface    $localeRepository    Locale repository
-     * @param CurrencyRepositoryInterface  $currencyRepository  Currency repository
-     * @param EventDispatcherInterface     $eventDispatcher     Event dispatcher
-     * @param array                        $classes             Product, product value and price classes
+     * @param AttributeRepositoryInterface       $attributeRepository Attribute repository
+     * @param FamilyRepositoryInterface          $familyRepository    Family repository
+     * @param ChannelRepositoryInterface         $channelRepository   Channel repository
+     * @param LocaleRepositoryInterface          $localeRepository    Locale repository
+     * @param CurrencyRepositoryInterface        $currencyRepository  Currency repository
+     * @param AssociationTypeRepositoryInterface $assocTypeRepository Association type repository
+     * @param EventDispatcherInterface           $eventDispatcher     Event dispatcher
+     * @param array                              $classes             Model classes
      */
     public function __construct(
         AttributeRepositoryInterface $attributeRepository,
@@ -65,18 +73,21 @@ class ProductBuilder implements ProductBuilderInterface
         ChannelRepositoryInterface $channelRepository,
         LocaleRepositoryInterface $localeRepository,
         CurrencyRepositoryInterface $currencyRepository,
+        AssociationTypeRepositoryInterface $assocTypeRepository,
         EventDispatcherInterface $eventDispatcher,
         array $classes
     ) {
         $this->attributeRepository = $attributeRepository;
-        $this->familyRepository = $familyRepository;
-        $this->channelRepository  = $channelRepository;
-        $this->localeRepository   = $localeRepository;
-        $this->currencyRepository = $currencyRepository;
-        $this->eventDispatcher    = $eventDispatcher;
-        $this->productClass       = $classes['product'];
-        $this->productValueClass  = $classes['product_value'];
-        $this->productPriceClass  = $classes['product_price'];
+        $this->familyRepository    = $familyRepository;
+        $this->channelRepository   = $channelRepository;
+        $this->localeRepository    = $localeRepository;
+        $this->currencyRepository  = $currencyRepository;
+        $this->assocTypeRepository = $assocTypeRepository;
+        $this->eventDispatcher     = $eventDispatcher;
+        $this->productClass        = $classes['product'];
+        $this->productValueClass   = $classes['product_value'];
+        $this->productPriceClass   = $classes['product_price'];
+        $this->associationClass    = $classes['association'];
     }
 
     /**
@@ -126,6 +137,25 @@ class ProductBuilder implements ProductBuilderInterface
         }
 
         $this->addMissingPricesToProduct($product);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function addMissingAssociations(ProductInterface $product)
+    {
+        $missingAssocTypes = $this->assocTypeRepository->findMissingAssociationTypes($product);
+        if (!empty($missingAssocTypes)) {
+            foreach ($missingAssocTypes as $associationType) {
+                $association = new $this->associationClass();
+                $association->setAssociationType($associationType);
+                $product->addAssociation($association);
+            }
+        }
+
+        return $this;
     }
 
     /**